requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
aiofiles>=23.2.1
asyncio>=3.4.3
typing-extensions>=4.5.0
dataclasses>=0.6

# Database and Storage
neo4j>=5.14.0
neo4j-driver>=5.14.0
chromadb>=0.4.22
networkx>=3.1
PyYAML>=6.0.1

# AI and Machine Learning
openai>=1.3.0
sentence-transformers>=2.2.2
numpy>=1.24.3
scikit-learn>=1.3.2

# Web and API
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.25.2
aiohttp>=3.9.1
websockets>=12.0
python-telegram-bot>=20.4

# Pattern Matching and Analysis
rapidfuzz>=3.5.2
python-Levenshtein>=0.21.1

# Utilities
tqdm>=4.66.1
python-dateutil>=2.8.2
PyJWT>=2.8.0
cryptography>=41.0.7
loguru>=0.7.2

# Machine Learning Extensions
torch>=2.1.1
transformers>=4.36.0

# Autogen Dependencies
autogen_ext>=0.5.6
autogen_agentchat>=0.5.6
autogen_core>=0.5.6
playwright>=1.52.0  # Required for web automation via autogen_ext
